refactor(www): get github stars
